Refactor upload source
Some testing

Fix python deploys
