+++ cmd/meroxa/turbine/utils.go
@@ -1,15 +1,11 @@
 package turbine
 
 import (
-	"archive/tar"
 	"bytes"
-	"compress/gzip"
 	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
 	"os"
 	"os/exec"
 	"path"
@@ -17,9 +13,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/meroxa/cli/cmd/meroxa/global"
 	"github.com/meroxa/cli/log"
-	turbineGo "github.com/meroxa/turbine-go/deploy"
 )
 
 const (
@@ -30,7 +24,6 @@ const (
 	Python3    = "python3"
 	Ruby       = "ruby"
 
-	TurbineJSVersion  = "1.3.5"
 	isTrue            = "true"
 	AccountUUIDEnvVar = "MEROXA_ACCOUNT_UUID"
 
@@ -257,13 +250,26 @@ func GetGitBranch(appPath string) (string, error) {
 	return strings.TrimSpace(string(output)), nil
 }
 
-// SwitchToAppDirectory switches temporarily to the application's directory.
-func SwitchToAppDirectory(appPath string) (string, error) {
-	pwd, err := os.Getwd()
-	if err != nil {
-		return pwd, err
+func GetTurbineResponseFromOutput(output string) (string, error) {
+	r := regexp.MustCompile("turbine-response: ([^\n]*)")
+	match := r.FindStringSubmatch(output)
+	if match == nil || len(match) < 2 {
+		return "", fmt.Errorf("output is formatted unexpectedly")
 	}
-	return pwd, os.Chdir(appPath)
+	return match[1], nil
+}
+
+func RunCMD(ctx context.Context, logger log.Logger, cmd *exec.Cmd) error {
+	if err := cmd.Start(); err != nil {
+		logger.Errorf(ctx, err.Error())
+		return err
+	}
+
+	if err := cmd.Wait(); err != nil {
+		logger.Errorf(ctx, err.Error())
+		return err
+	}
+	return nil
 }
 
 // RunCmdWithErrorDetection checks exit codes and stderr for failures and logs on success.
@@ -293,277 +299,22 @@ func RunCmdWithErrorDetection(ctx context.Context, cmd *exec.Cmd, l log.Logger) 
 	return stdOutMsg, nil
 }
 
-// CreateTarAndZipFile creates a .tar.gz file from `src` on current directory.
-func createTarAndZipFile(src string, buf io.Writer) error {
-	// Grab the directory we care about (app's directory)
-	appDir := filepath.Base(src)
-
-	// Change to parent's app directory
-	pwd, err := SwitchToAppDirectory(filepath.Dir(src))
+// SwitchToAppDirectory switches temporarily to the application's directory.
+func SwitchToAppDirectory(appPath string) (string, error) {
+	pwd, err := os.Getwd()
 	if err != nil {
-		return err
+		return pwd, err
 	}
-
-	zipWriter := gzip.NewWriter(buf)
-	tarWriter := tar.NewWriter(zipWriter)
-
-	err = filepath.Walk(appDir, func(file string, fi os.FileInfo, err error) error {
-		if shouldSkipDir(fi) {
-			return filepath.SkipDir
-		}
-		header, err := tar.FileInfoHeader(fi, file)
-		if err != nil {
-			return err
-		}
-
-		header.Name = filepath.ToSlash(file)
-		if err := tarWriter.WriteHeader(header); err != nil { //nolint:govet
-			return err
-		}
-		if !fi.Mode().IsRegular() {
-			return nil
-		}
-		if !fi.IsDir() {
-			var data *os.File
-			data, err = os.Open(file)
-			defer func(data *os.File) {
-				err = data.Close()
-				if err != nil {
-					panic(err.Error())
-				}
-			}(data)
-			if err != nil {
-				return err
-			}
-			if _, err := io.Copy(tarWriter, data); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-
-	if err := tarWriter.Close(); err != nil {
-		return err
-	}
-	if err := zipWriter.Close(); err != nil {
-		return err
-	}
-
-	return os.Chdir(pwd)
+	return pwd, os.Chdir(appPath)
 }
 
-func shouldSkipDir(fi os.FileInfo) bool {
-	if !fi.IsDir() {
-		return false
-	}
+func CleanupDockerfile(logger log.Logger, path string) {
+	logger.StartSpinner("\t", fmt.Sprintf("Removing Dockerfile created for your application in %s...", path))
+	// We clean up Dockerfile as last step
 
-	switch fi.Name() {
-	case ".git", "fixtures", "node_modules":
-		return true
-	}
-
-	return false
-}
-
-func UploadSource(ctx context.Context, logger log.Logger, language, appPath, appName, url string) error {
-	var err error
-
-	if language == GoLang || language == JavaScript {
-		logger.StartSpinner("\t", fmt.Sprintf("Creating Dockerfile before uploading source in %s", appPath))
-
-		if language == GoLang {
-			err = turbineGo.CreateDockerfile(appName, appPath)
-		}
-
-		if language == JavaScript {
-			err = createJavascriptDockerfile(ctx, appPath)
-		}
-
-		if err != nil {
-			return err
-		}
-		defer func() {
-			logger.StartSpinner("\t", fmt.Sprintf("Removing Dockerfile created for your application in %s...", appPath))
-			// We clean up Dockerfile as last step
-			err = os.Remove(filepath.Join(appPath, "Dockerfile"))
-			if err != nil {
-				logger.StopSpinnerWithStatus(fmt.Sprintf("Unable to remove Dockerfile: %v", err), log.Failed)
-			} else {
-				logger.StopSpinnerWithStatus("Dockerfile removed", log.Successful)
-			}
-		}()
-		logger.StopSpinnerWithStatus("Dockerfile created", log.Successful)
-	}
-
-	dFile := fmt.Sprintf("turbine-%s.tar.gz", appName)
-
-	var buf bytes.Buffer
-	err = createTarAndZipFile(appPath, &buf)
-	if err != nil {
-		return err
-	}
-
-	logger.StartSpinner("\t", fmt.Sprintf("Creating %q in %q to upload to our build service...", appPath, dFile))
-
-	fileToWrite, err := os.OpenFile(dFile, os.O_CREATE|os.O_RDWR, os.FileMode(0777)) //nolint:gomnd
-	defer func(fileToWrite *os.File) {
-		err = fileToWrite.Close()
-		if err != nil {
-			panic(err.Error())
-		}
-
-		// remove .tar.gz file
-		logger.StartSpinner("\t", fmt.Sprintf("Removing %q...", dFile))
-		removeErr := os.Remove(dFile)
-		if removeErr != nil {
-			logger.StopSpinnerWithStatus(fmt.Sprintf("\t Something went wrong trying to remove %q", dFile), log.Failed)
-		} else {
-			logger.StopSpinnerWithStatus(fmt.Sprintf("Removed %q", dFile), log.Successful)
-		}
-
-		if language == Python {
-			cleanUpPythonTempBuildLocation(ctx, logger, appPath)
-		}
-	}(fileToWrite)
-	if err != nil {
-		return err
-	}
-	if _, err = io.Copy(fileToWrite, &buf); err != nil {
-		return err
-	}
-	logger.StopSpinnerWithStatus(fmt.Sprintf("%q successfully created in %q", dFile, appPath), log.Successful)
-
-	return uploadFile(ctx, logger, dFile, url)
-}
-
-func uploadFile(ctx context.Context, logger log.Logger, filePath, url string) error {
-	logger.StartSpinner("\t", "Uploading source...")
-
-	var clientErr error
-	var res *http.Response
-	client := &http.Client{}
-
-	retries := 3
-	for retries > 0 {
-		fh, err := os.Open(filePath)
-		if err != nil {
-			logger.StopSpinnerWithStatus("\t Failed to open source file", log.Failed)
-			return err
-		}
-		defer func(fh *os.File) {
-			fh.Close()
-		}(fh)
-
-		req, err := http.NewRequestWithContext(ctx, "PUT", url, fh)
-		if err != nil {
-			logger.StopSpinnerWithStatus("\t Failed to make new request", log.Failed)
-			return err
-		}
-
-		fi, err := fh.Stat()
-		if err != nil {
-			logger.StopSpinnerWithStatus("\t Failed to stat source file", log.Failed)
-			return err
-		}
-		req.Header.Set("Accept", "*/*")
-		req.Header.Set("Content-Type", "multipart/form-data")
-		req.Header.Set("Accept-Encoding", "gzip, deflate, br")
-		req.Header.Set("Connection", "keep-alive")
-
-		req.ContentLength = fi.Size()
-
-		res, clientErr = client.Do(req)
-		if clientErr != nil || (res.StatusCode < http.StatusOK || res.StatusCode >= http.StatusMultipleChoices) {
-			retries--
-		} else {
-			break
-		}
-	}
-
-	if res != nil && res.Body != nil {
-		defer res.Body.Close()
-	}
-	if clientErr != nil || (res.StatusCode < http.StatusOK || res.StatusCode >= http.StatusMultipleChoices) {
-		logger.StopSpinnerWithStatus("\t Failed to upload build source file", log.Failed)
-		if clientErr == nil {
-			clientErr = fmt.Errorf("upload failed: %s", res.Status)
-		}
-		return clientErr
-	}
-
-	logger.StopSpinnerWithStatus("Source uploaded", log.Successful)
-	return nil
-}
-
-func RunTurbineJS(ctx context.Context, params ...string) (cmd *exec.Cmd) {
-	args := getTurbineJSBinary(params)
-	return executeTurbineJSCommand(ctx, args)
-}
-
-func getTurbineJSBinary(params []string) []string {
-	shouldUseLocalTurbineJS := global.GetLocalTurbineJSSetting()
-	turbineJSBinary := fmt.Sprintf("@meroxa/turbine-js-cli@%s", TurbineJSVersion)
-	if shouldUseLocalTurbineJS == isTrue {
-		turbineJSBinary = "turbine-js-cli"
-	}
-	args := []string{"npx", "--yes", turbineJSBinary}
-	args = append(args, params...)
-	return args
-}
-
-func executeTurbineJSCommand(ctx context.Context, params []string) *exec.Cmd {
-	return exec.CommandContext(ctx, params[0], params[1:]...) //nolint:gosec
-}
-
-func createJavascriptDockerfile(ctx context.Context, appPath string) error {
-	cmd := RunTurbineJS(ctx, "clibuild", appPath)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("unable to create Dockerfile at %s; %s", appPath, string(output))
-	}
-
-	return err
-}
-
-// cleanUpPythonTempBuildLocation removes any artifacts in the temporary directory.
-func cleanUpPythonTempBuildLocation(ctx context.Context, logger log.Logger, appPath string) {
-	logger.StartSpinner("\t", fmt.Sprintf("Removing artifacts from building the Python Application at %s...", appPath))
-
-	cmd := exec.CommandContext(ctx, "turbine-py", "cliclean", appPath)
-	output, err := cmd.CombinedOutput()
-
-	if err != nil {
-		logger.StopSpinnerWithStatus(fmt.Sprintf("\t Failed to clean up artifacts at %s: %v %s", appPath, err, output), log.Failed)
+	if err := os.Remove(filepath.Join(path, "Dockerfile")); err != nil {
+		logger.StopSpinnerWithStatus(fmt.Sprintf("Unable to remove Dockerfile: %v", err), log.Failed)
 	} else {
-		logger.StopSpinnerWithStatus("Removed artifacts from building", log.Successful)
+		logger.StopSpinnerWithStatus("Dockerfile removed", log.Successful)
 	}
-
-	if err != nil {
-		fmt.Printf("unable to clean up Meroxa Application at %s; %s", appPath, string(output))
-	}
-}
-
-func GetTurbineResponseFromOutput(output string) (string, error) {
-	r := regexp.MustCompile("turbine-response: ([^\n]*)")
-	match := r.FindStringSubmatch(output)
-	if match == nil || len(match) < 2 {
-		return "", fmt.Errorf("output is formatted unexpectedly")
-	}
-	return match[1], nil
-}
-
-func RunCMD(ctx context.Context, logger log.Logger, cmd *exec.Cmd) error {
-	if err := cmd.Start(); err != nil {
-		logger.Errorf(ctx, err.Error())
-		return err
-	}
-
-	if err := cmd.Wait(); err != nil {
-		logger.Errorf(ctx, err.Error())
-		return err
-	}
-	return nil
 }
